+++ lib/environment_config.rb
@@ -1,9 +1,22 @@
 # frozen_string_literal: true
 
 require 'environment_config/builder'
+require 'environment_config/type_fetcher_methods'
 
 class EnvironmentConfig
+  include TypeFetcherMethods
+
   class << self
+    # Accepts the same block as `load` does, but only validates
+    # the presence and type of environment variables. Does not return
+    # a configuration.
+    def ensure(&block)
+      load(&block)
+      nil
+    end
+
+    # Loads a configuration from environment variables as specified
+    # by the block given (using the environment config DSL)
     def load(**options)
       builder = Builder.new(**options)
       yield builder