+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.1.0
+
+* Add `EnvironmentConfig.ensure`, which is a more expressive way to
+  check the presence of environment variables without using them
+* Add `EnvironmentConfig.fetch_*` methods that will fetch a single
+  environment variable expecting the given type (e.g. `fetch_string`).
+* Improve error readout for unparsable integer
+
 ## 1.0.1
 
 * Make method private that should never have been public